+++ src/cal3d/renderer.h
@@ -16,7 +16,7 @@ class CalModel;
 class CalSubmesh;
 
 namespace CalRenderer {
-    CAL3D_API int getMapCount(const CalSubmesh* submesh);
+    CAL3D_API size_t getMapCount(const CalSubmesh* submesh);
     CAL3D_API Cal::UserData* getMaterialUserData(const CalSubmesh* submesh);
     CAL3D_API bool textureCoordinatesForMapValid(const CalSubmesh* submesh, int mapId);
 };
